Fix phpcs: extract test helper classes into separate files
